frontend/requirement.txt
fastapi==0.70.0
requests==2.26.0
uvicorn==0.15.0
opencensus-ext-azure==1.1.0
opencensus-ext-requests==0.7.6